chore(frontend): move Plan/logic to React layer + extend CI guard
components/Plan/logic/{title,issue,title.test}.ts were framework-neutral
helpers (no Vue specifics, pure pinia+router plumbing) with zero Vue
callers — only 7 React files and 1 React test mock referenced them.
The Plan/ directory existed only to hold this logic/ subdir.

- Moved 3 files to src/react/lib/plan/
- Updated 7 React imports and 1 vi.mock path
- Removed the now-empty src/components/Plan/ directory
- Extended no-react-to-vue-imports.test.ts to ban @/components/Plan/
  from the React layer
